pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = "API for managing Blog, comments with AI"
authors = ["Denys <https://github.com/AnyoneClown>"]
license = "Apache License"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
fastapi = {extras = ["standart"], version = "^0.115.2"}
uvicorn = "^0.32.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
google-generativeai = "^0.8.3"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.36"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
